changed double to int

+++ ElgrosLib/Factories/ProductFactory.cs
@@ -13,7 +13,7 @@ namespace ElgrosLib.Factories
         /// </summary>
         /// <param name="name"></param>
         /// <returns></returns>
-        public static Product CreateProduct(int id, string name, string description, double price, double quantity, string photoUrl, int categoryId, int subCategoryId)
+        public static Product CreateProduct(int id, string name, string description, double price, int quantity, string photoUrl, int categoryId, int subCategoryId)
         {
             return new Product(id, name, description, price, quantity, photoUrl, categoryId, subCategoryId);
         }
